Factor out AVX2/FMA composition functions
